scripts/curriculum-scraper/requirements.txt
# Curriculum Scraper Dependencies
# Install with: pip install -r requirements.txt

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0  # For JavaScript-rendered content

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# Data validation
jsonschema>=4.20.0
pydantic>=2.5.0

# Utilities
tqdm>=4.66.0  # Progress bars
python-slugify>=8.0.0
pandas>=2.1.0  # Data analysis

